# ApiBarberShop/ApiBarberShop/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ApiBarberShop, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Barberos/GetBarberos": {
      "get": {
        "tags": [
          "Barberos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Barbero"
              }
            }
          }
        }
      }
    },
    "/Barberos/GetBarberos{id}": {
      "get": {
        "tags": [
          "Barberos"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Barbero"
            }
          }
        }
      }
    },
    "/Barberos/PutBarberos{id}": {
      "put": {
        "tags": [
          "Barberos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Barbero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Barberos/PostBarberos": {
      "post": {
        "tags": [
          "Barberos"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Barbero"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Barbero"
            }
          }
        }
      }
    },
    "/Barberos/Delete{id}": {
      "delete": {
        "tags": [
          "Barberos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Citas/GetCitas": {
      "get": {
        "tags": [
          "Citas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cita"
              }
            }
          }
        }
      }
    },
    "/Citas/GetCita{id}": {
      "get": {
        "tags": [
          "Citas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cita"
            }
          }
        }
      }
    },
    "/Citas/PutCitas{id}": {
      "put": {
        "tags": [
          "Citas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cita"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Citas/PostCitas": {
      "post": {
        "tags": [
          "Citas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cita"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cita"
            }
          }
        }
      }
    },
    "/Citas/Delete{id}": {
      "delete": {
        "tags": [
          "Citas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Clientes/GetClientes": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cliente"
              }
            }
          }
        }
      }
    },
    "/Clientes/GetClientes{id}": {
      "get": {
        "tags": [
          "Clientes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        }
      }
    },
    "/Clientes/PutClientes{id}": {
      "put": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Clientes/PostClientes": {
      "post": {
        "tags": [
          "Clientes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        }
      }
    },
    "/Clientes/Delete{id}": {
      "delete": {
        "tags": [
          "Clientes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Servicios/GetServicios": {
      "get": {
        "tags": [
          "Servicios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Servicio"
              }
            }
          }
        }
      }
    },
    "/Servicios/GetServicios{id}": {
      "get": {
        "tags": [
          "Servicios"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Servicio"
            }
          }
        }
      }
    },
    "/Servicios/PutServicios{id}": {
      "put": {
        "tags": [
          "Servicios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Servicio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Servicios/PostServicios": {
      "post": {
        "tags": [
          "Servicios"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Servicio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Servicio"
            }
          }
        }
      }
    },
    "/Servicios/Delete{id}": {
      "delete": {
        "tags": [
          "Servicios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Barbero": {
      "type": "object",
      "properties": {
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "fechaCreacion": {
          "format": "date-time",
          "type": "string"
        },
        "fechaModificacion": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "cita": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Cita"
          },
          "readOnly": true
        }
      }
    },
    "Cita": {
      "type": "object",
      "properties": {
        "citaId": {
          "format": "int32",
          "type": "integer"
        },
        "servicioId": {
          "format": "int32",
          "type": "integer"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "mensaje": {
          "type": "string"
        },
        "fechaCreacion": {
          "format": "date-time",
          "type": "string"
        },
        "fechaModificacion": {
          "format": "date-time",
          "type": "string"
        },
        "usuarioCreacionId": {
          "format": "int32",
          "type": "integer"
        },
        "usuarioModificacionId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "barbero": {
          "$ref": "#/definitions/Barbero"
        },
        "cliente": {
          "$ref": "#/definitions/Cliente"
        },
        "servicio": {
          "$ref": "#/definitions/Servicio"
        }
      }
    },
    "Cliente": {
      "type": "object",
      "properties": {
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "apellido": {
          "type": "string"
        },
        "celular": {
          "type": "string"
        },
        "fechaNacimiento": {
          "format": "date-time",
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "fechaCreacion": {
          "format": "date-time",
          "type": "string"
        },
        "fechaModificacion": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "cita": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Cita"
          },
          "readOnly": true
        }
      }
    },
    "Servicio": {
      "type": "object",
      "properties": {
        "servicioId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "fechaCreacion": {
          "format": "date-time",
          "type": "string"
        },
        "fechaModificacion": {
          "format": "date-time",
          "type": "string"
        },
        "usuarioCreacionId": {
          "format": "int32",
          "type": "integer"
        },
        "usuarioModificacionId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "cita": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Cita"
          },
          "readOnly": true
        }
      }
    }
  }
}